[routing] Rename bike sharing station module
Rename the bicycle rental station lookup module to match its responsibility.
