+++ libs/routing/CMakeLists.txt
@@ -5,8 +5,8 @@ set(SRC
   absent_regions_finder.hpp
   async_router.cpp
   async_router.hpp
-  bike_sharing_heuristic.cpp
-  bike_sharing_heuristic.hpp
+  bicycle_rental.cpp
+  bicycle_rental.hpp
   base/astar_algorithm.hpp
   base/astar_progress.cpp
   base/astar_progress.hpp
